Fix bug preventing users from logging in
When the UserAuthenticationForm is initialized from the view, it isn't passed
the request. This leaves the next parameter blank if the user opened the login
directly, which confused django and prevented the users from logging in.

Just adding a "/" as a default value for the hidden value will make sure it
isn't empty if no request is given, thus fixing this bug.

It also checks if next="" is explicitly given and corrects that to "/".