Added ruby script to correct the translation stuff
It looks through the po file, possibly switching german strings in the source
files for english ones and then switching the german msgids with the english
msgstrs.

This will make all strings in the source text english and create a proper german
translation, which should leave us with a proper i18n/l10n state.